got all logic into splash shard removed maincontent and replaced with fragment space
